A crbug (https://www.crbug.com/988153) about
ui::AXNode::ComputeLastUnignoredChildRecursive is fixed and needs to be
merge into release branch 77.

The merge of those two CLs has been approved:
1. https://chromium-review.googlesource.com/c/chromium/src/+/1737870
2. https://chromium-review.googlesource.com/c/chromium/src/+/1739629

I cherry picked them into branch 3865 (77). There were merge conflicts
in the modified unit test. I resolved the conflicts, built chromium and
built & run the tests and they all passed.

IMO, the merge is safe, but it can't hurt to have a second opinion on
this.

(cherry picked from commit db7a7f3c9fdb66111cf3fd2f9e02b40c53ff6ef5 [formerly 74a89b88ea5a34e2a8c43f14a7dc6c76cc62f14c])
(cherry picked from commit f79f928226be44fa4f8da4aeb9b596f99c2c1c0e [formerly 46ebf53ad51dd084bc7b04964838acf85217201b])

Bug: 988153
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1757176
Cr-Commit-Position: refs/branch-heads/3865@{#429}
Cr-Branched-From: 4e0147fd4ab868c78ff6bb1ef42f86871cde1461 [formerly 0cdcc6158160790658d1f033d3db873603250124]-refs/heads/master@{#681094}

Former-commit-id: ddf8a9e14c0125bb8364c7a603a885989c081973

+++ ui/accessibility/ax_node.cc
@@ -764,7 +764,7 @@ AXNode* AXNode::ComputeLastUnignoredChildRecursive() const {
   if (children().size() == 0)
     return nullptr;
 
-  for (size_t i = children().size() - 1; i >= 0; --i) {
+  for (int i = static_cast<int>(children().size()) - 1; i >= 0; --i) {
     AXNode* child = children_[i];
     if (!child->data().HasState(ax::mojom::State::kIgnored))
       return child;
